another case of generics registration going wrong on Container.Configure()
