+++ src/StructureMap.Testing/StructureMap.Testing.csproj
@@ -222,7 +222,7 @@
     <Compile Include="Bugs\closed_type_generic_is_in_get_all.cs" />
     <Compile Include="Bugs\CloseOpenGenericsWithSomeSpecifics.cs" />
     <Compile Include="Bugs\ConnectImplementationsToTypesClosing_is_wonky_in_Registry_added_by_Configure.cs" />
-    <Compile Include="Bugs\Container_Configure_crashes_on_generics_contravariance.cs" />
+    <Compile Include="Bugs\container_configuration_with_generics_inheritance.cs" />
     <Compile Include="Bugs\Defensive_check_for_public_ctors.cs" />
     <Compile Include="Bugs\do_not_allow_policy_changes_in_nested_container_issue_284.cs" />
     <Compile Include="Bugs\EnumerableShouldGetAllValuesTester.cs" />
